+++ tools/auth/auth_test.go
@@ -189,4 +189,13 @@ func TestNewProviderByName(t *testing.T) {
 	if _, ok := p.(*auth.Instagram); !ok {
 		t.Error("Expected to be instance of *auth.Instagram")
 	}
+
+	// vk
+	p, err = auth.NewProviderByName(auth.NameVK)
+	if err != nil {
+		t.Errorf("Expected nil, got error %v", err)
+	}
+	if _, ok := p.(*auth.VK); !ok {
+		t.Error("Expected to be instance of *auth.VK")
+	}
 }